36 corrected 0 charges and changes to Db structure. Trade duration no records seconds in decimals

+++ core/positions/position.py
@@ -41,7 +41,7 @@ class Position:
 
     net_pnl: float = 0.0
     
-    duration_seconds: int = 0
+    duration_seconds: float = 0
     
     initial_stop_loss: float = 0.0
